backend/requirements.txt
# FastAPI
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1

# Pydantic
pydantic==2.10.4
pydantic-settings==2.7.1

# Google Ads
google-ads==30.0.0

# Meta Ads
facebook-business

# Google Analytics 4
google-analytics-data==0.18.14

# Shopify
httpx==0.28.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==44.0.0

# Task queue (Phase 3)
celery[redis]==5.4.0
redis==5.2.1

# Utilities
python-dotenv==1.0.1

# Scheduler
apscheduler==3.10.4

# Export
openpyxl==3.1.5

